\name{tbasscpp-package}
\alias{tbasscpp-package}
\alias{tbasscpp}
\docType{package}
\title{
\packageTitle{tbasscpp}
}
\description{
\packageDescription{tbasscpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tbasscpp}
\packageIndices{tbasscpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tbasscpp}

Maintainer: \packageMaintainer{tbasscpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
